requirements.txt
numpy
pandas
plotly
pyro-ppl
torch
ollama
seaborn
matplotlib
scipy
scikit-learn
mlflow
tensorboard